+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Use `on-finished` to determine request status
+
 0.2.0 / 2014-09-03
 ==================